neuroglancer.annotations.precompted: Rearranged some function definitions
