+++ neuclease/misc/neuroglancer/annotations/precomputed/_spatial.py
@@ -3,9 +3,10 @@ from itertools import chain
 from typing import NamedTuple
 
 import numpy as np
-from numba import njit, guvectorize
+from numba import njit
 from numba.typed import List
 
+from .compressed_morton import compressed_morton_code, reverse_morton_code, _compressed_morton_code_pairwise
 from ._write_buffers import _write_buffers
 from ._util import _encode_uint64_series, _geometry_cols, _ndindex_array
 
@@ -76,6 +77,107 @@ def _assign_spatial_chunks(df, coord_space, annotation_type, bounds, num_levels,
     return df, gridspec
 
 
+def _define_spatial_grids(bounds, coord_space, num_levels: int) -> GridSpec:
+    """
+    Compute suitable chunk shapes and grid shapes for each level 
+    of the spatial index, following the guidelines from the spec[1]:
+
+        Typically the grid_shape for level 0 should be a vector of all 1
+        (with chunk_size equal to upper_bound - lower_bound), and each component
+        of chunk_size of each successively level should be either equal to, or half of,
+        the corresponding component of the prior level chunk_size, whichever results
+        in a more spatially isotropic chunk.
+
+    [1]: https://github.com/google/neuroglancer/blob/master/src/datasource/precomputed/annotations.md#spatial-index
+
+    Args:
+        bounds:
+            np.ndarray, shape (2, D)
+            lower and upper bounds of the union of all annotations
+        num_levels:
+            The number of spatial index levels. Must be at least 1.
+
+    Returns:
+        chunk_shapes, grid_shapes
+    """
+    # Level 0 chunk shape and grid shape -- just one chunk.
+    bounds = np.asarray(bounds, np.float64)
+
+    # We want roughly isotropic chunks in physical units, so we'll multiply
+    # by the coordinate scales and then divide the scales out at the end.
+    chunk_shape = (bounds[1] - bounds[0]) * coord_space.scales
+    grid_shape = np.ones_like(chunk_shape, dtype=np.uint64)
+
+    chunk_shapes = [chunk_shape]
+    grid_shapes = [grid_shape]
+
+    for level in range(1, num_levels):
+        chunk_shape = chunk_shape.copy()
+        grid_shape = grid_shape.copy()
+
+        max_dim = np.argmax(chunk_shape)
+        target_width = chunk_shape[max_dim] / 2
+
+        for dim, dim_width in enumerate(chunk_shape):
+            if dim == max_dim:
+                # Always split across the widest dimension.
+                chunk_shape[dim] = target_width
+                grid_shape[dim] *= 2
+            elif (dim_width / target_width) > 1.5:
+                # Split across this dimension to make it more isotropic.
+                chunk_shape[dim] = dim_width / 2
+                grid_shape[dim] *= 2
+            else:
+                # Splitting would make it less isotropic,
+                # so leave this dimension unsplit.
+                chunk_shape[dim] = dim_width
+                grid_shape[dim] *= 1
+
+        chunk_shapes.append(chunk_shape)
+        grid_shapes.append(grid_shape)
+
+    # Convert from physical units back to coordinate units.
+    chunk_shapes = np.array(chunk_shapes, dtype=np.float64) / coord_space.scales
+    grid_shapes = np.array(grid_shapes, dtype=np.uint64)
+
+    return GridSpec(chunk_shapes, grid_shapes)
+
+
+def _compute_target_annotations_per_level(num_annotations, gridspec, target_chunk_limit: int):
+    """
+    Compute the TOTAL number of annotations at each level of the spatial index.
+    The target_chunk_limit is how many annotations we aim to place in each chunk
+    (regardless of the level).
+    
+    Since the spatial annotations are not distributed uniformly in space,
+    we will likely end up undershooting and overshooting the target for various
+    chunks within a level.
+
+    Furthermore, since the number of annotations passed in here is based on the
+    table BEFORE duplicating annotations which span multiple chunks, the number
+    of annotations at each level will eventually be more than what is returned here,
+    after the appropriate duplications.
+
+    Returns:
+        np.ndarray, shape (num_levels,)
+    """
+    num_levels = len(gridspec.grid_shapes)
+    chunk_counts_by_level = np.prod(gridspec.grid_shapes, axis=1)
+    annotation_counts = chunk_counts_by_level * target_chunk_limit
+    
+    # Clamp to total number of annotations remaining after earlier levels
+    for level in range(num_levels - 1):
+        annotation_counts[level] = min(
+            annotation_counts[level],
+            num_annotations - sum(annotation_counts[:level])
+        )
+
+    # Last level gets all remaining annotations, if any.
+    annotation_counts[-1] = num_annotations - sum(annotation_counts[:-1])
+
+    return annotation_counts
+
+
 def _assign_spatial_chunks_for_points(df, geometry_cols, bounds, gridspec):
     coord_names = geometry_cols[0]
     grid_indices = (df[[*coord_names]] - bounds[0]) // gridspec.chunk_shapes[df['level']]
@@ -90,6 +192,7 @@ def _assign_spatial_chunks_for_points(df, geometry_cols, bounds, gridspec):
         gridspec.grid_shapes[df['level']][:, ::-1]
     )
     
+    # FIXME: remove this debug code or put it somewhere else.
     # import pyarrow.feather as feather
     # df[[f'grid_{c}' for c in coord_names]] = grid_indices
     # df[[f'{c}a' for c in coord_names]] =  grid_indices * gridspec.chunk_shapes[df['level']] + bounds[0]
@@ -299,107 +402,6 @@ def _line_chunk_overlap(point_a, point_b, grid_origin, cell_shape, grid_index):
     return max_t >= min_t
 
 
-def _define_spatial_grids(bounds, coord_space, num_levels: int) -> GridSpec:
-    """
-    Compute suitable chunk shapes and grid shapes for each level 
-    of the spatial index, following the guidelines from the spec[1]:
-
-        Typically the grid_shape for level 0 should be a vector of all 1
-        (with chunk_size equal to upper_bound - lower_bound), and each component
-        of chunk_size of each successively level should be either equal to, or half of,
-        the corresponding component of the prior level chunk_size, whichever results
-        in a more spatially isotropic chunk.
-
-    [1]: https://github.com/google/neuroglancer/blob/master/src/datasource/precomputed/annotations.md#spatial-index
-
-    Args:
-        bounds:
-            np.ndarray, shape (2, D)
-            lower and upper bounds of the union of all annotations
-        num_levels:
-            The number of spatial index levels. Must be at least 1.
-
-    Returns:
-        chunk_shapes, grid_shapes
-    """
-    # Level 0 chunk shape and grid shape -- just one chunk.
-    bounds = np.asarray(bounds, np.float64)
-
-    # We want roughly isotropic chunks in physical units, so we'll multiply
-    # by the coordinate scales and then divide the scales out at the end.
-    chunk_shape = (bounds[1] - bounds[0]) * coord_space.scales
-    grid_shape = np.ones_like(chunk_shape, dtype=np.uint64)
-
-    chunk_shapes = [chunk_shape]
-    grid_shapes = [grid_shape]
-
-    for level in range(1, num_levels):
-        chunk_shape = chunk_shape.copy()
-        grid_shape = grid_shape.copy()
-
-        max_dim = np.argmax(chunk_shape)
-        target_width = chunk_shape[max_dim] / 2
-
-        for dim, dim_width in enumerate(chunk_shape):
-            if dim == max_dim:
-                # Always split across the widest dimension.
-                chunk_shape[dim] = target_width
-                grid_shape[dim] *= 2
-            elif (dim_width / target_width) > 1.5:
-                # Split across this dimension to make it more isotropic.
-                chunk_shape[dim] = dim_width / 2
-                grid_shape[dim] *= 2
-            else:
-                # Splitting would make it less isotropic,
-                # so leave this dimension unsplit.
-                chunk_shape[dim] = dim_width
-                grid_shape[dim] *= 1
-
-        chunk_shapes.append(chunk_shape)
-        grid_shapes.append(grid_shape)
-
-    # Convert from physical units back to coordinate units.
-    chunk_shapes = np.array(chunk_shapes, dtype=np.float64) / coord_space.scales
-    grid_shapes = np.array(grid_shapes, dtype=np.uint64)
-
-    return GridSpec(chunk_shapes, grid_shapes)
-
-
-def _compute_target_annotations_per_level(num_annotations, gridspec, target_chunk_limit: int):
-    """
-    Compute the TOTAL number of annotations at each level of the spatial index.
-    The target_chunk_limit is how many annotations we aim to place in each chunk
-    (regardless of the level).
-    
-    Since the spatial annotations are not distributed uniformly in space,
-    we will likely end up undershooting and overshooting the target for various
-    chunks within a level.
-
-    Furthermore, since the number of annotations passed in here is based on the
-    table BEFORE duplicating annotations which span multiple chunks, the number
-    of annotations at each level will eventually be more than what is returned here,
-    after the appropriate duplications.
-
-    Returns:
-        np.ndarray, shape (num_levels,)
-    """
-    num_levels = len(gridspec.grid_shapes)
-    chunk_counts_by_level = np.prod(gridspec.grid_shapes, axis=1)
-    annotation_counts = chunk_counts_by_level * target_chunk_limit
-    
-    # Clamp to total number of annotations remaining after earlier levels
-    for level in range(num_levels - 1):
-        annotation_counts[level] = min(
-            annotation_counts[level],
-            num_annotations - sum(annotation_counts[:level])
-        )
-
-    # Last level gets all remaining annotations, if any.
-    annotation_counts[-1] = num_annotations - sum(annotation_counts[:-1])
-
-    return annotation_counts
-
-
 def __write_assigned_annotations_by_spatial_chunk(df, gridspec, output_dir, write_sharded):
     """
     Write the spatial index, given a dataframe in which the 'level'
@@ -444,139 +446,3 @@ def __write_assigned_annotations_by_spatial_chunk(df, gridspec, output_dir, writ
         metadata.append(level_metadata)
 
     return metadata
-
-
-def compressed_morton_code(grid_coord_c_order, grid_shape_c_order):
-    """
-    Return the compressed morton code[1] for a grid coordinate
-    that resides in a grid with the given shape.
-
-    [1]: https://github.com/google/neuroglancer/blob/master/src/datasource/precomputed/volume.md#compressed-morton-code
-
-    This implementation uses numba.guvectorize().
-    Multiple coordinates and/or grid shapes can be provided simultaneously,
-    and they will be broadcasted together to produce an output of the appropriate shape.
-
-    Important:
-    
-        This function expects the grid coordinate to be in C order,
-        i.e. the fastest-changing stride is listed last.
-        For typical 3D arrays, C-order indexing is [z, y, x],
-        so pass your grid coordinates here in [z, y, x] order, too.
-
-    Examples:
-
-        In [10]: compressed_morton_code(
-            ...:     [1,5,3],
-            ...:     [3, 16, 5]
-            ...: )
-        Out[10]: array(143, dtype=uint64)
-
-        In [11]: compressed_morton_code(
-            ...:     [[1,5,3], [2, 3, 4]],
-            ...:     [3, 16, 5]
-            ...: )
-        Out[11]: array([143, 114], dtype=uint64)
-
-        In [12]: compressed_morton_code(
-            ...:     [1,5,3],
-            ...:     [[3, 16, 5], [6, 32, 10]]
-            ...: )
-        Out[12]: array([143, 143], dtype=uint64)
-
-        In [13]: compressed_morton_code(
-            ...:     [[1,5,3], [2, 3, 4]],
-            ...:     [[3, 16, 5], [6, 32, 10]]
-            ...: )
-        Out[13]: array([143, 114], dtype=uint64)
-    """
-    grid_coord = np.asarray(grid_coord_c_order, np.uint64)
-    grid_shape = np.asarray(grid_shape_c_order, np.uint64)
-    if (grid_coord >= grid_shape).any():
-        raise ValueError(f'grid_coord contains out-of-bounds values relative to grid_shape')
-
-    axis_bits = np.ceil(np.log2(grid_shape)).astype(np.int8)
-    output_shape = np.broadcast_shapes(grid_shape.shape, grid_coord.shape)[:-1]
-    output_code = np.zeros(output_shape, dtype=np.uint64)
-    __compressed_morton_code(grid_coord, axis_bits, output_code)
-    return output_code
-
-
-@guvectorize('(d),(d)->()', nopython=True)
-def __compressed_morton_code(grid_coord, axis_bits, result):
-    D = len(axis_bits)
-    curr_axis_pos = np.zeros(D, dtype=np.uint64)
-    curr_axis = 0
-    output_pos = np.uint64(0)
-    output_code = np.uint64(0)
-
-    for _ in range(D * 64):
-        curr_axis = (curr_axis - 1) % D
-        if curr_axis_pos[curr_axis] >= axis_bits[curr_axis]:
-            continue
-
-        input_pos = curr_axis_pos[curr_axis]
-        output_code |= ((grid_coord[curr_axis] >> input_pos) & np.uint64(1)) << output_pos
-        output_pos += np.uint64(1)
-        curr_axis_pos[curr_axis] += 1
-
-    # With guvectorize, we treat a scalar output as if it had a single element.
-    # https://numba.readthedocs.io/en/stable/user/vectorize.html#scalar-return-values
-    result[0] = output_code
-
-
-@njit
-def _compressed_morton_code_pairwise(grid_coord_c_order, grid_shape_c_order):
-    """
-    Same as compressed_morton_code(), but for when both grid_coord
-    and grid_shape are known to have shape ndim 1 or 2 and both arguments
-    have dtype np.uint64, which allows us to easily predict the output
-    shape without needing to call np.broadcast_shapes().
-
-    Therefore this whole function can be wrapped with @njit,
-    and can be called from within other jit-compiled functions.
-    """
-    grid_coord = grid_coord_c_order
-    grid_shape = grid_shape_c_order
-
-    axis_bits = np.ceil(np.log2(grid_shape)).astype(np.int8)
-    if grid_coord.ndim == 1:
-        output_code = np.zeros(1, dtype=np.uint64)
-        __compressed_morton_code(grid_coord, axis_bits, output_code)
-        return output_code[0]
-    else:
-        output_code = np.zeros(len(grid_coord), dtype=np.uint64)
-        __compressed_morton_code(grid_coord, axis_bits, output_code)
-        return output_code
-
-
-def reverse_morton_code(morton_code, grid_shape_c_order):
-    morton_code = np.asarray(morton_code, np.uint64)
-    grid_shape = np.asarray(grid_shape_c_order, np.uint64)
-
-    D = grid_shape.shape[-1]
-    axis_bits = np.ceil(np.log2(grid_shape)).astype(np.int8)
-    output_shape = np.broadcast_shapes(morton_code.shape, grid_shape.shape[:-1]) + (D,)
-    output_grid_coord = np.zeros(output_shape, dtype=np.uint64)
-    __reverse_morton_code(morton_code, axis_bits, output_grid_coord)
-    return output_grid_coord
-
-
-@guvectorize('(),(d)->(d)', nopython=True)
-def __reverse_morton_code(morton_code, axis_bits, grid_coord_c_order):
-    D = len(axis_bits)
-    curr_axis_pos = np.zeros(D, dtype=np.uint64)
-    curr_axis = 0
-    
-    input_pos = np.uint64(0)
-    grid_coord_c_order[:] = np.uint64(0)
-
-    for _ in range(D * 64):
-        curr_axis = (curr_axis - 1) % D
-        if curr_axis_pos[curr_axis] >= axis_bits[curr_axis]:
-            continue
-
-        output_pos = curr_axis_pos[curr_axis]
-        grid_coord_c_order[curr_axis] |= ((morton_code >> input_pos) & np.uint64(1)) << output_pos
-        input_pos += np.uint64(1)
-        curr_axis_pos[curr_axis] += 1
